refactor: move MessageType from thread.type to message.type
