+++ client/src/component/organism/ThreadDetail/ThreadDetail.tsx
@@ -1,11 +1,11 @@
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 import { RootState } from '@store'
 import A from '@atom'
 import O from '@organism'
 import { TextType } from '@atom/Text'
-import { MessageType } from '@type/thread.type'
 import {
+  MessageType,
   CreateMessageRequestType,
   MessageWithThreadIdType,
 } from '@type/message.type'
