+++ client/src/type/message.type.ts
@@ -1,9 +1,33 @@
-import {
-  CreateThreadRequestType,
-  MessageType,
-  GetThreadResponseType,
-} from './thread.type'
+import { UserType } from './user.type'
 import { ResponseType } from './response.type'
+import { CreateThreadRequestType, GetThreadResponseType } from './thread.type'
+
+interface FileType {
+  id: number
+  url: string
+  type: string
+  createdAt?: string
+  updatedAt?: string
+}
+
+interface ReactionType {
+  id: number
+  content: string
+  createdAt?: string
+  updatedAt?: string
+  User: UserType
+}
+
+export interface MessageType {
+  id: number
+  content: string
+  isHead?: false
+  createdAt: string
+  updatedAt: string
+  User: UserType
+  File: FileType[]
+  Reactions: ReactionType[]
+}
 
 export interface UpdateMessageRequestType extends CreateThreadRequestType {
   messageId: number
